refactor(core): extract pure pool/RAM sizing helpers
Move cgroup/RAM probes and pool/thread budgeting into inference/sizing
so load math is free of ONNX sessions. Engine keeps thin wrappers for
existing unit-test call sites; behavior is unchanged.
